refactor(sync_handlers): name v1 explicitly to make rebases less painful
