+++ Cargo.toml
@@ -10,9 +10,9 @@ members = [
     "crates/merkle-tree",
     "crates/p2p",
     "crates/p2p_bootstrap",
-    "crates/p2p_proto",
     "crates/p2p_proto_derive",
     "crates/p2p_proto_v0",
+    "crates/p2p_proto_v1",
     "crates/pathfinder",
     "crates/retry",
     "crates/rpc",
