+++ crates/p2p_proto_v1/build.rs
