+++ crates/p2p_proto_v1/proto/snapshot.proto
