+++ crates/p2p_proto_v1/src/state.rs
