+++ crates/pathfinder/Cargo.toml
@@ -13,7 +13,7 @@ path = "src/lib.rs"
 [features]
 tokio-console = ["console-subscriber", "tokio/tracing"]
 rpc-full-serde = []
-p2p = ["dep:base64", "dep:p2p", "dep:p2p_proto", "dep:p2p_proto_v0", "dep:zeroize"]
+p2p = ["dep:base64", "dep:p2p", "dep:p2p_proto_v0", "dep:p2p_proto_v1", "dep:zeroize"]
 
 [dependencies]
 anyhow = { workspace = true }
@@ -30,8 +30,8 @@ metrics = { workspace = true }
 metrics-exporter-prometheus = "0.11.0"
 num_cpus = "1.16.0"
 p2p = { path = "../p2p", optional = true }
-p2p_proto = { path = "../p2p_proto", optional = true }
 p2p_proto_v0 = { path = "../p2p_proto_v0", optional = true }
+p2p_proto_v1 = { path = "../p2p_proto_v1", optional = true }
 pathfinder-common = { path = "../common" }
 pathfinder-compiler = { path = "../compiler" }
 pathfinder-ethereum = { path = "../ethereum" }
